prediction_server/requirements.txt
Flask==2.3.3
gunicorn==21.2.0
scikit-learn==1.2.2
pandas==1.5.3
joblib==1.3.2
google-cloud-storage==2.14.0